server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.121.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-genai==1.65.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
numpy==2.4.2
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.5
starlette==0.49.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
websockets==16.0
